app/requirements.txt
fastapi
python-dotenv
uvicorn
langchain-text-splitters
tiktoken
sentence-transformers
qdrant-client
openai